+++ libraries/LTR390_RT/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2024-11-26
+- add **float getLUX()**
+- add **float getUVIndex()**
+- add **float getIntegrationTime()**
+- add **float getMeasurementTime()**
+- add **uint8_t getGainFactor()**
+- update readme.md, tables.
+
 ## [0.1.1] - 2024-05-09
 - fix URL's